Docker: Use assembled app from first stage of docker build.
Before the "locally" assembled app (i.e. outside of docker build stage) was used.
That lead to failing docker builds, if no local maven built was done
This is also unexpected, as the first stage of the docker build does its own maven build, which has was not used by seconds stage.