[JENKINS-54424] - Remove Jenkins Test Harness Dependency & replace by local code
This change removes dependency on the Jenkins Test Harness framework. JUnit is still used, but there is a separate ticket for that.

* Many classes were copied from JTH, but it can be optimized later
* Irrelevant code in JenkinsRule and other classes was removed

There is still a lot of optimizations/refactorings to be done (e.g. prevent WAR exploder from running at all, remove interfaces), but at least we keep the engine separate from JTH going forward